+++ PAWpy/__init__.py
@@ -23,7 +23,7 @@ from PAWpy.version_requirements import (
     is_supported,
 )
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 __all__ = [
     "PAWService",